#299 move recreating a container out of DockerClient
The sequence and everything it needs is its own class now, so DockerClient is
back to being a thin wrapper over docker-java.
